s/getcord/spacedentist (#204)
Summary:

Ownership was transfered a while ago, but links were never updated.
Noticed by clicking on the README link -- that was the only one which
was actually broken. As for the others, they were correctly redirecting
to the new URLs, but I've updated them for correctness.

Test Plan:

Visit a couple of those links, they work.